requirements.txt
fastapi
uvicorn
python-multipart
llama-index
llama-index-vector-stores-postgres
llama-index-llms-gemini
pgvector
psycopg2-binary
sqlalchemy
asyncpg
pydantic
pypdf
python-docx
pandas
huggingface
dotenv
passlib
python-jose[cryptography]
llama-index-embeddings-huggingface
passlib[bcrypt]